requirements.txt
djangorestframework
Django==2.2.9
django-allauth==0.41.0
django-celery==3.3.1
django-crispy-forms==1.8.1
django-extensions==2.2.6
django-guardian==2.2.0
django-service-objects==0.6.0
django-silk==4.0.0
flake8==3.7.9
isort==4.3.21
mysqlclient==1.4.6
Pillow==7.0.0
redis==2.10.6
twilio==6.35.5
stripe==2.42.0


